# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: LaunchKit AI API
  version: 1.0.0
  description: |
    LaunchKit AI API enables creators to automatically transform their existing content 
    into monetizable online courses with integrated payment processing.
    
    ## Features
    - Content extraction from Notion and other sources
    - AI-powered course generation and structuring
    - Stripe payment integration and webhook handling
    - Course delivery and access management
    
    ## Base URL
    - Development: `http://localhost:3001/api/v1`
    - Production: `https://api.launchkit.ai/v1`
  contact:
    name: LaunchKit AI Support
    url: https://launchkit.ai/support
    email: support@launchkit.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.launchkit.ai/v1
    description: Production server
components:
  schemas:
...
    Course:
      type: object
      required:
        - id
        - slug
        - title
        - ownerId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the course
          example: "clp456def789ghi012"
        slug:
          type: string
          description: URL-friendly unique identifier
          example: "intro-to-react-hooks"
        title:
          type: string
          description: Course title
          example: "Introduction to React Hooks"
        description:
          type: string
          nullable: true
          description: Course description
          example: "Learn modern React development with hooks"
        ownerId:
          type: string
          description: ID of the user who owns this course
          example: "clp123abc456def789"
        createdAt:
          type: string
          format: date-time
          description: When the course was created
          example: "2024-01-15T14:22:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the course was last updated
          example: "2024-01-16T09:15:00Z"
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
          description: Lessons in this course
    Lesson:
      type: object
      required:
        - id
        - courseId
        - title
        - order
      properties:
        id:
          type: string
          description: Unique identifier for the lesson
          example: "clp789ghi012jkl345"
        courseId:
          type: string
          description: ID of the course this lesson belongs to
          example: "clp456def789ghi012"
        title:
          type: string
          description: Lesson title
          example: "useState Hook Basics"
        content:
          type: string
          nullable: true
          description: Lesson content in markdown format
          example: "# useState Hook\n\nThe useState hook allows you to..."
        order:
          type: integer
          minimum: 0
          default: 0
          description: Display order of the lesson within the course
          example: 1
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input data"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "email"
                  message:
                    type: string
                    example: "Invalid email format"
            requestId:
              type: string
              description: Unique request identifier for debugging
              example: "req_1234567890"
paths:
...
  /courses:
    get:
      summary: List user's courses
      description: Retrieve all courses created by a specific user
      tags:
        - Course Management
      parameters:
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
          description: ID of the course owner
          example: "clp123abc456def789"
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new course
      description: Create a new course with lessons for the specified owner
      tags:
        - Course Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ownerId
                - title
                - description
                - lessons
              properties:
                ownerId:
                  type: string
                  description: ID of the course owner
                  example: "clp123abc456def789"
                title:
                  type: string
                  description: Course title
                  example: "Introduction to JavaScript"
                description:
                  type: string
                  description: Course description
                  example: "Learn the fundamentals of JavaScript programming"
                lessons:
                  type: array
                  items:
                    type: object
                    required:
                      - title
                      - content
                    properties:
                      title:
                        type: string
                        description: Lesson title
                        example: "Variables and Data Types"
                      content:
                        type: string
                        description: Lesson content
                        example: "In this lesson, we'll learn about..."
                      order:
                        type: integer
                        description: Lesson order (optional, defaults to 0)
                        example: 1
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...